src/test/java/dev/petproject/controller/CategoryControllerTest.java
package dev.petproject.controller;

import dev.petproject.auth.JwtUtils;
import dev.petproject.repository.ProductRepository;
import dev.petproject.repository.TokenRepository;
import dev.petproject.service.CategoryService;
import dev.petproject.service.UserService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.MockitoAnnotations;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@WebMvcTest(CategoryController.class)
class CategoryControllerTest {

    @MockBean
    UserService userService;

    @MockBean
    JwtUtils jwtUtils;

    @MockBean
    TokenRepository tokenRepository;


    @MockBean
    private CategoryService categoryService;


    private MockMvc mockMvc;

    @MockBean
    private ProductRepository productRepository;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
        mockMvc = MockMvcBuilders.standaloneSetup(new CategoryController(categoryService)).build();
    }

    @Test
    void createCategory() throws Exception {
        mockMvc.perform(post("/save"))
                .andExpect(status().is3xxRedirection())
                .andExpect(redirectedUrl("/products/add?success"));
    }

    @Test
    void shouldOpenPageWithFormToCreateCategory() throws Exception {
        mockMvc.perform(get("/create"))
                .andExpect(status().isOk())
                .andExpect(model().attributeExists("category"))
                .andExpect(view().name("createCategory"));
    }
}